+++ packages/object-field-tree/src/index.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-implicit-globals */
 import type { DataFn1, DataFn2, DataFn3, DataFn4, DataFn5 } from './type/data-fn.type';
 import type { CombineReturn1, CombineReturn2, CombineReturn3, CombineReturn4, CombineReturn5 } from './type/return.type';
 import type { Enum } from '@rnw-community/shared';
@@ -35,7 +36,7 @@ export function combine<D, T1 extends Enum, T2 extends Enum, T3 extends Enum, T4
 ): CombineReturn5<D, T1, T2, T3, T4, T5>;
 
 // TODO: Introduce non-recursive optimized solution
-// eslint-disable-next-line func-style,@typescript-eslint/no-explicit-any
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function combine(dataFn: (...keys: any) => any, ...objects: any[]): any {
     const result = {};