+++ packages/react-native-payments-example-lib/src/component/request-options-form.tsx
@@ -27,6 +27,7 @@ export const RequestOptionsForm = ({ setResponse, setError }: RequestOptionsForm
 
     const clearErrorAndResponse = (): void => {
         setError('');
+        // eslint-disable-next-line no-undefined
         setResponse(undefined);
     };