+++ packages/wdio/src/component$/get-exteded-component$/get-extended-component$.spec.ts
@@ -6,7 +6,7 @@ import { ParentComponent$SelectorsMock } from '../mocks/parent-component$-select
 
 describe('getExtendedComponent$', () => {
     it('should use extended parent component$ testID selector functions', async () => {
-        // eslint-disable-next-line no-magic-numbers
+         
         expect.assertions(6);
 
         const component = new ExtendedComponent$Mock();