+++ packages/wdio/src/wdio.mock.ts
@@ -2,7 +2,7 @@
 import type { ChainablePromiseArray, ChainablePromiseElement } from 'webdriverio';
 import { jest } from '@jest/globals';
 
-// eslint-disable-next-line @typescript-eslint/consistent-type-assertions
+ 
 const mockWdioElement = {
     elementId: 'test-element-id',
     click: jest.fn(() => Promise.resolve(void 0)),