+++ packages/react-native-payments-example/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.7.0](https://github.com/rnw-community/rnw-community/compare/v1.6.2...v1.7.0) (2025-10-24)
+
+**Note:** Version bump only for package @rnw-community/react-native-payments-example
+
 ## [1.6.2](https://github.com/rnw-community/rnw-community/compare/v1.6.1...v1.6.2) (2025-08-01)
 
 **Note:** Version bump only for package @rnw-community/react-native-payments-example